+++ plugins/org.eclipse.app4mc.amalthea.model/src-gen/org/eclipse/app4mc/amalthea/model/impl/ContinuousHistogramImpl.java
@@ -131,14 +131,14 @@ public class ContinuousHistogramImpl extends AmaltheaExtendedEObjectImpl impleme
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public double getAverage() {
+	public Double getAverage() {
 		double _xblockexpression = (double) 0;
 		{
 			final Function1<ContinuousHistogramEntry, Double> _function = new Function1<ContinuousHistogramEntry, Double>() {
 				public Double apply(final ContinuousHistogramEntry it) {
-					double _average = it.getInterval().getAverage();
+					Double _average = it.getInterval().getAverage();
 					long _occurrences = it.getOccurrences();
-					return Double.valueOf((_average * _occurrences));
+					return Double.valueOf(((_average).doubleValue() * _occurrences));
 				}
 			};
 			final Function2<Double, Double, Double> _function_1 = new Function2<Double, Double, Double>() {
@@ -160,7 +160,7 @@ public class ContinuousHistogramImpl extends AmaltheaExtendedEObjectImpl impleme
 			final Long count = IterableExtensions.<Long>reduce(XcoreEListExtensions.<ContinuousHistogramEntry, Long>map(this.getEntries(), _function_2), _function_3);
 			_xblockexpression = DoubleExtensions.operator_divide(sum, count);
 		}
-		return _xblockexpression;
+		return Double.valueOf(_xblockexpression);
 	}
 
 	/**
