+++ plugins/org.eclipse.app4mc.amalthea.model/src-gen/org/eclipse/app4mc/amalthea/model/impl/DiscreteConstantImpl.java
@@ -117,8 +117,8 @@ public class DiscreteConstantImpl extends AmaltheaExtendedEObjectImpl implements
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public double getAverage() {
-		return (this.getValue()).longValue();
+	public Double getAverage() {
+		return Double.valueOf(this.getValue().doubleValue());
 	}
 
 	/**
